Use structure expected by autoloader instead of custom structure
This commit undoes the prior structure of the gem for clarity purposes.
