+++ .autoupdate/postupdate
@@ -2,7 +2,7 @@
 
 # This script is called by our weekly dependency update job in Jenkins after updating Ruby and other deps
 
-# Switch to Ruby 3.1 for Globus::Client (3.0 is default in Jenkinsfile)
+# Switch to Ruby 3.1 for GlobusClient (3.0 is default in Jenkinsfile)
 rvm use 3.1.2@globus_client --create &&
     gem install bundler &&
     bundle install --gemfile Gemfile
