--- lib/globus/client/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Globus
-  class Client
-    VERSION = "0.3.2"
-  end
-end
