+++ lib/globus_client/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class GlobusClient
+  VERSION = "0.3.2"
+end
